реализовали вход и регистрацию, а также возможность удалять только свои заметки

+++ notes.controller.js
@@ -1,28 +1,41 @@
-const chalk = require('chalk') // подключаем пакет для украшения консоли
+const chalk = require('chalk'); // подключаем пакет для украшения консоли
 // подключаем модель Note для получения возможности чтения и записи в базу данных
-const Note = require('./models/Note')
+const Note = require('./models/Note');
 
-async function addNote(title) {
+async function addNote(title, owner) {
 	// const notes = require('./db.json')     // обращение к базе данных делается по другому
 	// const notes = Buffer.from(buffer).toString('utf-8')
 	// чтобы не прерывать процесс программы NodeJS считывая файлы складывает результат в буффер
-	await Note.create({ title })
-	console.log(chalk.bgGreen('Note was added!'))
+	await Note.create({ title, owner });
+	console.log(chalk.bgGreen('Note was added!'));
 }
 
 async function getNotes() {
-	const notes = await Note.find()
-	return Array.isArray(notes) ? notes : []
+	const notes = await Note.find();
+	return Array.isArray(notes) ? notes : [];
 }
 
-async function updateNote(newTitle, id) {
-	await Note.updateOne({ _id: id }, { title: newTitle })
-	console.log(chalk.bgGreen('Note was updated!'))
+async function removeNote(id, owner) {
+	const result = await Note.deleteOne({ _id: id, owner });
+
+	if (result.matchedCount === 0) {
+		throw new Error('No note to delete!');
+	}
+
+	console.log(chalk.bgGreen('Note was deleted!'));
 }
 
-async function removeNote(id) {
-	await Note.deleteOne({ _id: id })
-	console.log(chalk.bgGreen('Note was deleted!'))
+async function updateNote(noteData, owner) {
+	const result = await Note.updateOne(
+		{ _id: noteData.id, owner },
+		{ title: noteData.newTitle },
+	);
+
+	if (result.matchedCount === 0) {
+		throw new Error('No note to edit!');
+	}
+
+	console.log(chalk.bgGreen('Note was updated!'));
 }
 
 // экспортируем эти функции
@@ -31,4 +44,4 @@ module.exports = {
 	getNotes,
 	removeNote,
 	updateNote,
-}
+};
